add Makefile, move configs to internal
