+++ cmd/client/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"fmt"
 	"github.com/pkg/errors"
-	"github.com/putalexey/goph-keeper/cmd/client/config"
+	"github.com/putalexey/goph-keeper/internal/client/config"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 	"log"
